refactor: rename langdetect to language_detector to be more clear and better match the other modules in the package
