+++ simplemma/language_detector.py
